mcp/requirements.txt
mcp>=1.0.0
sentence-transformers>=3.0.1
faiss-cpu>=1.8.0
PyMuPDF>=1.24.4
requests>=2.32.2
PyYAML>=6.0.1
beautifulsoup4>=4.12.3
rich>=13.7.1
